Rewrite `MergeParameters` to work with TS 4.9
Per https://github.com/microsoft/TypeScript/pull/50831 , the existing
implementation of `MergeParameters` broke with TS 49 due to a change
in how "optional" and "undefined" arguments get interpreted

Anders Hjelsberg himself supplied a new implementation that's actually
much simpler, but only works with TS 4.7 and greater.

Normally this would require shipping two entirely different sets of
types, as we already do for TS <4.1. However, there's a trick we used
with RTK where we:

- Create a folder and put two different files with different impls of
  the same type inside
- Add an index file and re-export one of those
- Add a file named `package.dist.json` containing `typesVersions` that
  points to both of those `.d.ts` files
- Copy that `package.dist.json` over to `dist/some/package.json` during
  the actual build/publish step

That way during dev TS just imports the type as normal, but the built
version sees that `some/package.json`, sees `typesVersions`, finds the
right `.d.ts` file, and imports the right implementation for itself.
