+++ .eslintrc
@@ -45,7 +45,7 @@
         "@typescript-eslint/no-explicit-any": "off",
         "@typescript-eslint/no-unused-vars": "off",
         "@typescript-eslint/no-non-null-assertion": "off",
-        "@typescript-eslint/no-shadow": ["error"],
+        "@typescript-eslint/no-shadow": ["off"],
         "@typescript-eslint/no-use-before-define": ["error"],
         "@typescript-eslint/ban-types": "off",
         "prefer-rest-params": "off",
