+++ .github/workflows/build-and-test-types.yml
@@ -62,14 +62,23 @@ jobs:
       - name: Install deps
         run: yarn install
 
-      - name: Install TypeScript ${{ matrix.ts }}
-        run: yarn add typescript@${{ matrix.ts }}
-
+      # Build with the actual TS version in the repo
       - name: Pack
         run: yarn build && yarn pack
 
       - name: Install build artifact
         run: yarn add ./package.tgz
+
+      # Then install the specific version to test against
+      - name: Install TypeScript ${{ matrix.ts }}
+        run: yarn add --dev typescript@${{ matrix.ts }}
+
+      - name: 'Remove source to ensure packaged types are used'
+        run: rm -rf src
+
+        # Remove config line that points "reselect" to the `src` folder,
+        # so that the typetest will use the installed version instead
+      - run: sed -i -e /@remap-prod-remove-line/d ./typescript_test/tsconfig.json
 
       - name: Test types
         run: |
