+++ src/versionedTypes/index.ts
@@ -1,0 +1,1 @@
+export { MergeParameters } from './ts47-mergeParameters'
